inferentia/requirements.txt
streamlit>=1.36.0
flask>=3.0.3
fastapi>=0.111.1
uvicorn>=0.30.1
gunicorn>=23.0.0
locust>=2.29.1
tiktoken>=0.6.0
blobfile>=2.1.1
fairscale>=0.4.13
fire
ray[serve]>=2.32.0
vllm>=0.6.0
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
neuronx-cc==2.* --extra-index-url https://pip.repos.neuron.amazonaws.com
neuronx-cc-stubs --extra-index-url https://pip.repos.neuron.amazonaws.com
neuronx-distributed --extra-index-url https://pip.repos.neuron.amazonaws.com
torch-neuronx --extra-index-url https://pip.repos.neuron.amazonaws.com
libneuronxla --extra-index-url https://pip.repos.neuron.amazonaws.com
transformers_neuronx --extra-index-url https://pip.repos.neuron.amazonaws.com
aws-neuronx-runtime-discovery --extra-index-url https://pip.repos.neuron.amazonaws.com
optimum-neuron @git+https://github.com/huggingface/optimum-neuron.git
transformers>=4.43.2 #@git+https://github.com/huggingface/transformers.git
